Merged PR 21982: 4.7.0-RC1 release notes
4.7.0 release candidate 1.

We're adopting a new strategy for connector releases. In an effort to align better with Magento's own release cycle, we'll be releasing new features in our bundled releases, as far as possible. We'll be releasing chunks of release candidate code as we go - of which this is the first - and tagging the develop branch.

Interim patch releases (tagged on `master`) will be for bug fixes and minor improvements only.

Related work items: #116808