Merged PR 22389: Add extra check for empty $response
## What's being changed

We're adding an extra check for an empty `$response` when making an API call to Engagement Cloud for surveys and forms.

## Why it's being changed

Our check for `isset($response->message)` won't log anything if `$response` is not an object. Albeit unlikely, we'd want to know if this is the case.

Related work items: #117778