Merged PR 22685: Set nullable property on last_subscribed_at
## What's being changed

We're setting the `nullable` property on the `last_subscribed_at` column, in the install schema.

## Why it's being changed

We introduced the `last_subscribed_at` column to the `email_contact` table in v3.2.2. In the upgrade schema, we specify that the field is `'nullable' => true` but we don't in InstallSchema. We're adding it here for consistency.

## How to review / test this change

- Create a clean Magento instance
- Before installing, set `Dotdigitalgroup_Email` to 3.2.2 for schema and data in `setup_module`
- Run `setup:install`
- Confirm that `last_subscribed_at` in `email_contact` indicates 'Allow null'

Related work items: #119415