Merged PR 23300: Add return value to getProductCategories()
## What's being changed

We're adding a return value to the getProductCategories() method, called when adding custom extension attributes for web behaviour tracking.

## Why it's being changed

Merchants observed that WBT data in Engagement Cloud did not display `product_categories` even when products in the store had categories.

Unfortunately when writing this code for 4.3.6, a change was made after initial testing that removed the return value from the method (unless an exception was thrown). We overlooked the missing data in subsequent tests.

## How to review / test this change

- Repeat tests for this original PR: https://dev.azure.com/dotdigital/ec/_git/ec-core-magento2-extension/pullrequest/17782

Related work items: #122104