Merged PR 22914: Include review campaign automation in review sync CLI command
## What's being changed

The code that generates review campaigns, if review automation is enabled, is being moved to the main `Model\Sync\Review` class.

## Why it's being changed

This code previously sat in a script run only via the "review and wishlist sync" cron, and was not being run via the `dotdigital:sync Review` CLI command.

## How to review / test this change

- Enable review automation
- Place an order
- Run `php bin/magento dotdigital:sync Review`
- Assuming your order fits the conditions configured in dotdigital > Automation > Review Settings, you should see an 'Order Review' campaign in your `email_campaign` table

Related work items: #120899