Merged PR 24814: Fix multiple store registrations for same subscriber between cron runs
## What's being changed

We've added a store filter when reloading the contact collection in subscriber sync.

## Why it's being changed

In subscriber sync, we:

- get customers who are subscribers
- get guests who are subscribers
- extract guests with sales data into their own separate cohort
- reload a collection of contacts containing the customers, and the non-sales guests

When reloading, we were only searching by email. So in the rare case where you subscribed on website 1, then subscribed on website 2, before the cron had a chance to run, you would only be synced to the address book of the first website.  This has now been fixed.

This PR also updates the subscriber sync by removing any proxy loading of collections via `$this->getCollection()` in a model.

## How to review / test this change

- Set up a Magento with two websites
- Ensure crons are off
- Sign up the same email address to each website
- Run subscriber sync
- Confirm that two importer rows are created and the contact is included in both csv files

Related work items: #127092