Merged PR 24439: Fetch store name, store view name and website name datafields in guest sync
## What's being changed

We added we store name, store view and website name columns in guest sync

## Why it's being changed

Before that we were using the $website so we were fetching the website name in store view name column in guests csv. In addition now we decided to add all these three columns.

## How to review / test this change

- Place an order as a guest
- Run guest sync
- Make sure that we have WEBSITE_NAME, STORE_NAME (and the manually chosen one for Store Name label) with the correct values in guest csv

Related work items: #125814