Merged PR 25888: Cache coupon block response
## What's being changed

We've added caching to the coupon block, in the form of a class property.

## Why it's being changed

Some merchants have experienced issues when generating coupons, repeatedly triggering the 'Already processed' exception in `DotdigitalCouponRequestProcessor`.

Evidently in some setups, where we check for the response in the EDC controller:
```
if (strlen($this->_view->getLayout()->getOutput()) < 10)
```
the block is being requested and rendered a second time. In this case the `$couponGeneratorStatus` property is already set, and we throw an exception. Expected behaviour is that calling `getLayout()->getOutput()` should return the value for `ddg.coupon.container` from the `_renderElementCache` array in `Magento\Framework\View\Layout`.

Adding the variable in the block ensures that this is returned _in case_ `_renderElementCache['ddg.coupon.container']` is not set, and the `Layout` class decides to render the element again.

## How to review / test this change

- Hit a coupon EDC

## Notes

We may want to scale this out to other EDCs.

Related work items: #131205