Merged PR 26400: Update transactional email observer
## What's being changed

The check for a host value is now more robust in this observer. Fixes BUNDLE-2871.

## Why it's being changed

The host check assumed that a value would always be set - however now that Email Templates are being saved as part of the same form, it is possible for `$groups['ddg_transactional']['fields']['host']['value']` to not be set at point of save. This fix resolves the error message.

## How to review / test this change

- Go to configure email templates at store view level
- Have Transactional SMS Host set to inherit
- Set a new template and save
- Ensure new mapping is saved as expected
- Test unchecking 'inherit' on the Host input and saving 'Please Select' (you will get the 'Please select a host' error)

Related work items: #132909