Merged PR 26533: Remove reset before set in invalid creds account test
## What's being changed

We're removing the initial `ResetDotmailerAccountData` command in this test.

## Why it's being changed

The intention was to reset in advance to ensure we start from a blank state - however this fails because the username and password fields are not displayed if the Yes/No select is set to 'No'.

## How to review / test this change

```
vendor/bin/mftf run:test ConnectWithValidApiCredentialsTest -r
```

Related work items: #133274