Merged PR 26557: Remove scrollTo step in cart price rule test
## What's being changed

This PR removes a step from a core action group, when our extension is enabled.

## Why it's being changed

Our theory is that the `scrollToCustomerGroupsControl` step doesn't work because the 'dotdigital Coupon URL Builder' block is affecting the scroll target. This is pretty tenuous because scrollTo is supposed to scroll to the element top. So it could be the customer groups select not loading in time. Either way, scrollTo isn't necessary because the select is visible without scrolling.

## How to review / test this change
```
vendor/bin/mftf run:test AdminDisplayingCartAndCatalogPriceRulesGridsTest -r
```

Related work items: #133333