Merged PR 26962: Fix scope for Disable Customer Success and Disable Newsletter Success
## What's being changed

The two plugins we use to prevent sending email notifications for new customers and new subscribers, depending on configured values.

## Why it's being changed

Configuration > Admin > Disable Customer Success / Disable Newsletter Success are website-scoped, but we were fetching the values in store scope.

## How to review / test this change

- Set Configuration > Admin > Disable Customer Success to Yes
- Create a new customer either from the admin or the front end
- Confirm that an email notification is not sent

- Set Configuration > Admin > Disable Newsletter Success to Yes
- Subscribe to the newsletter
- Confirm that an email notification is not sent

## Notes

In the CustomerEmailNotificationPlugin, the conditional statement and method:
```
if (!$storeId) {
    $storeId = $this->getWebsiteStoreId($customer, $sendemailStoreId);
}
```

is copied directly from the method being plugged (`newAccount()`). It's hard to see how `$storeId` could be null, but I suppose it makes sense to preserve this.

Related work items: #135003