Merged PR 27533: cURL requests: only set a message property if there is one
## What's being changed

The logic added to Rest.php in this PR: https://dev.azure.com/dotdigital/ec/_git/ec-core-magento2-extension/pullrequest/27060.

## Why it's being changed

Even if `$this->getCurlError()` was `false`, we were creating a `$response->message`, which our scripts check for when deciding if there was an error with the response.

Rather than change all the `isset()` checks, we now confirm an error exists before assigning it to the `message`.

This fixes misleading error messages both in the logs and in the customer account (when unsubscribing).

## How to review / test this change

- Run subscriber sync
- Confirm no error log lines are output to connector.log if no suppressed contacts are found.

Related work items: #136672