Merged PR 26895: Support global account sharing
## What's being changed

This PR brings support for merchants running Magento with **Customers > Customer Configuration > Account Sharing Options > Share Customer Accounts** set to 'Global'. It is now possible to have the flexibility of global customer access to all sites in a Magento network, combined with the separation of customer groups by website in Engagement Cloud.

Most of the changes here apply to the places in the connector that handle rows in our `email_contact` table. Broadly, we are relying less on store and website values in the `customer_entity` table, and more on either the current scope (for store front observers and plugins) or our own table's values.

To summarise the biggest areas of change:

- A `CustomerLogin` observer now adds new rows in `email_contact` for customers logging into other sites in the same network.
- The install / migrate process now contains additional migrations (if global account sharing is enabled).
- Customer email address changes are now copied to the other 'accounts' we have for that email address and customer id.
- Sales data field values are scoped to the store ids of the website we are syncing.
- Wishlist observers and wishlist sync work together to send wishlist items to an EC account based on their store id.

## Why it's being changed

Before these adjustments, merchants opting to share customer accounts globally i.e. across multiple websites in a network, could not integrate properly with Engagement Cloud (either with a single EC account or multiple EC accounts).

## Notes

There should be no change in behaviour for merchants with **Customers > Customer Configuration > Account Sharing Options > Share Customer Accounts** set to 'Per Website'.

Related work items: #134356, #132947, #133911, #133906, #133908, #133912, #134644, #134645, #134869, #134343, #136126, #136212