Merged PR 26462: Retrieve dashboard module list from plugins in sibling modules
## What's being changed

This PR moves a lot of logic to do with identifying sibling dotdigital module versions into their own respective modules. The core module now has a base method, `fetchActiveModules()`, and the other modules each have a plugin which adds their own bit of data to the list.

## Why it's being changed

This plugin approach is an improvement because it properly gives each module responsibility for displaying information about itself in the report dashboard.

## How to review / test this change

- Switch ALL four of the modules to branch `AM/132727-better`
- `setup:di:compile`
- Reload the dashboard report
- Check all modules are listed in the correct order
- Disable one module, does it disappear from the list after clearing cache? Do the others still list OK?

## Notes

The main issue with this is that it will require any installed sibling module to be the latest one, or otherwise Dashboard will be incomplete.

Related work items: #132727, #135666