Merged PR 27662: Use quote currency for line items in cartInsight data
Fetch quote currency for lineitems in cartInsight data

## What's being changed

In this PR we edited the price values of the quote items in CartInsight data.
We added also a new function to convert the prices based on the quote selected currency.

## Why it's being changed

In a multi currencies store we use to fetch only the price based on the base currency for some of the price values.

## How to review / test this change

In a multi currency store
- Drop an item to cart (you need to test it at least 2 times by selecting the base and non the base currency)
- Make sure that the price fields in cartInsight data are correct

Related work items: #132448