Merged PR 28525: Restrict stock update parent lookup to configurable type only
## What's being changed

ParentFinder.

## Why it's being changed

We have a method `getConfigurableParentsFromBunchOfProducts()` that we introduced in order to provide up to date stock figures for configurable parents. If a simple product is modified in certain contexts, and we're about to mark that product for sync in `email_catalog`, then this method finds the simple product's configurable parent, if it exists, and if it does we mark that one as unprocessed also.

We had an issue raised to us via Magento support whereby an admin user - with a role scope limited to one particular website - was unable to modify a product inside their own scope, if that product had been added as a child product to a bundle product that was _outside_ of their scope control.

The reason for the error was our `StockUpdatePlugin`. This is triggered at product save. We were looking up and loading parent products that were outside of the current user's scope.

The solution was to restrict `getConfigurableParentsFromBunchOfProducts()` to look up _only_ configurable products. This fixes the error as raised.

Note that for configurable products, restricted admin users cannot access simple children where the parent is outside their scope - this is core Magento behaviour and not influenced by our code. For bundle products, this change means that restricted admin users can view and edit simple products even if they are included as bundle options in bundle products that are outside of their scope.

## How to review / test this change

- Setup two websites (ex: "Main Website", "Japan website")
- Create a new admin user role, set Role Resource to "All", set Role Scopes to "Custom" and then select "Japan website".
- Create an admin account with this role
- Then using your own super user account, create a simple product A, and assign this product to both "Main Website" and "Japan website".
- Create a bundle product B only on "Main Website" then include product A as a child product.
- Now login to the admin panel as the Japan website-specific user then go to product A edit page and try to save the simple product A.
- This should succeed, you should not see the "More permissions are needed to view this item." error.

Related work items: #140106