Merged PR 29533: Do not set allow_non_subscribers values to 1 at install
## What's being changed

Our UpgradeData script.

## Why it's being changed

All settings for `allow_non_subscribers` (there are three switches for general sync, abandoned carts and reviews) should default to `0` as per config.xml.

However, prior to versions 2.5.0 and 2.5.1, allowing non-subscribers was the default behaviour of the module, and there were no switches or configuration for this. The steps in UpgradeData were added in order to preserve this behaviour for merchants already up and running on the older code.

The problem is that fresh installs would actually default to `1` - so to prevent this, we've added an extra condition in the script i.e. the setting of `allow_non_subscribers` will only happen for merchants on versions between 2.0.0 and 2.5.0/2.5.1.

## How to review / test this change

- Run a fresh install of our module (easiest way is to run an integration test with cleanup enabled)
- Do not see any settings for `allow_non_subscribers` set to `1` in `core_config_data`.
- Reset `setup_module` version to 2.0.0
- Run setup:upgrade
- Confirm that `allow_non_subscribers` set to `1` for all three keys

Related work items: #143275