Merged PR 29139: Convert install and upgrade scripts to declarative schema
## What's being changed

Our InstallSchema script has been replaced with a declarative schema definition and initial db_schema_whitelist file.

## Why it's being changed

This cleaner approach to handling schema changes was introduced in 2.3, and is not backwards compatible with 2.2. We have only just been able to end support for 2.2.

## How to review / test this change

- bin/magento setup:install --dry-run=1
- bin/magento setup:upgrade --dry-run=1
- Check log file at /var/log/dry-run-installation.log
- Run setup:install on a fresh environment
- Minutely compare each table in the new environment with tables in an old environment - has _anything_ changed? Check relations, indexes, definitions, comments - everything.

## Notes

InstallData is still in place to perform the initial migration. We experimented with replacing this with a patch but this isn't how patches work, they aren't designed to replace historical instructions.

UpgradeSchema and UpgradeData are still in place to handle incremental upgrades for merchants on older versions. Any modifications to either Schema or Data that need to be made _after_ this release will be made using patches.

Related work items: #102875