Merged PR 29985: Filter TransportPlugin error messages
## What's being changed

In this PR we filter the errors returned from TransportPlugin when we attempt to send a transactional email message and prevent the storing of system message flags for specific messages.

## Why it's being changed

If our plugin tries to send an email to an invalid email address, the EC SMTP relay might return the error message:
```
'Requested action not taken: mailbox unavailable'
```

In such a case, there is nothing the merchant can do, and it's not worth flagging as a system message. We will continue to log the error, and improve this log by indicating the specific email address that failed.

We may add additional 'excluded' errors to this filter over time.

## How to review / test this change

- Try to send an email to an invalid email address (via subscription customer creation etc)
- Make sure no new record is created in flag table

Related work items: #145162