Merged PR 30035:  Add shipping to cart insight data
## What's being changed

In this PR we added a new field to cart insight data for the quote shipping cost.

## Why it's being changed

Currently we provide the grand total in the data, but if shipping has been set on the quote, the calculation won't add up. If shipping is not set on the quote, we will send 0 (float).

## How to review / test this change

- Drop a cart let abandoned cart cron run and confirm that in cartInsight data, a new field "shipping" is created that displays separately the shipping cost.

Related work items: #135118