Merged PR 30289: Check if connector is enabled for the selected scope when automap datafields
## What's being changed

Currently we check if connector is enabled ONLY in default level so to decide if we should proceed with the data fields automap (when user clicks to run automap) or not. We changed that so to check if our connector is enabled in to at most of one of all the websites.

## Why it's being changed

There is a case that merchants have disabled the connector at default level but enabled only for one or few websites. In this case running automap data fields button doesn't work.

## How to review / test this change

- Disable connector in default level
- Enabled connector only to one website
- Try to automap datafields for the enabled website and confirm that happens as expected
- Confirm that you can't automap data fields for the websites the connector is disabled

Related work items: #144406