Merged PR 30268: In cart insight data, return item's product itself if sku not found
## What's being changed

In this PR we return item's product itself in cartInsight data if product failed to be loaded via sku or id depending if is configurable or not.

## Why it's being changed

In case that configurable products have custom options selected the sku is something like  COAT-123-EXTRA-1 EXTRA-2 EXTRA-3 etc (where EXTRA-1 EXTRA-2 EXTRA-3 is the options SKU's).

This sku is dynamically generated and doesn't exist in DB, as a result an exception thrown and line items are empty in CartinsightData collection.

With this PR, we
- get the dynamically generated sku
- As a unit price we get the price without the selected options
- As a sales price we get the price with options selected

## How to review / test this change

- Create a configurable product and add few custom options
- Drop the product in the cart
- Run AC cron
- Review cartinsight data and confirm that line items exist

Related work items: #139497