Merged PR 30345: Remove code from Zend Framework
## What's being changed

We're replacing some bits of code that either directly use the Zend Framework library, or work around its usage in older versions of Magento.

Specifically we've:
- replaced all usages of `\Zend_Date` with native PHP classes and constants
- removed handling of `\Zend_Mail` for SMTP

## Why it's being changed

ZF1 was updated to ZF2 in Magento 2.2. Then in 2.3.5, Zend was [migrated to](https://community.magento.com/t5/Magento-DevBlog/Migration-of-Zend-Framework-to-the-Laminas-Project/ba-p/443251) the Laminas project.

We might not be ready to require 2.3.5+ for using our connectors, so we can't yet replace all references to `\Zend_Class` with `\Laminas_Class`. However we can safely start the process by removing or replacing this older code.

## How to review / test this change

- Check one of the EDCs powered by the `Recommended` helper class. Are the date period settings translated and applied correctly?
- Send transactional email e.g. via newsletter subscribe.

Related work items: #93688