Merged PR 29888: Extend integration insight data to fetch the configuration values
## What's being changed

In this PR we extended the integration insight schema so to send all of the desired dotdigital configurations to Engagement Cloud. We will send the data only via a weekly cron, and not at upgrade/api save as previously.

## Why it's being changed

Our intention here is to have a greater knowledge of merchants' usage of our extensive module feature-set. As the core module evolves, this information will help us remain focussed on providing optimal utility.

## How to review / test this change

- Run the IntegrationInsights sync from the cli with:
```
bin/magento dotdigital:sync IntegrationInsights
```
- Check all keys and nested keys in the `configuration` array in Engagement Cloud
- Modify a value of one nested key, re-sync, and confirm the value is changed (test e.g. `null` to not null)

## Notes

Also in this PR we changed the record key. It changed from website url to website name because there is a possibility to override data in case that multiple websites share the same url.

Related work items: #144073